Implementing projects paypal payment.

+++ src/main/java/org/mql/crowddonating/business/IAssociationBusiness.java
@@ -40,7 +40,7 @@ public interface IAssociationBusiness extends IUserServices{
 	/* *** for signup *** */
 	void signup(Association association);
 	
-	Project deleteproject(int id);
+	Project deleteproject(long id);
 	String addProject(Project project);
 	String updateProject(Project project);
